Struct concurrent_hashmap::Options

pub struct Options<H> {
    pub capacity: usize,
    pub hasher_factory: H,
    pub concurrency: u16,
}

Options used when creating a hashmap.

Fields
capacity: usize

Number of mappings to preallocate space for.

The map will always grow as needed, but preallocating space can improve performance. This value applies to the entire map. By default, no space is preallocated.

hasher_factory: H

Factory for the hasher used for hashing keys.

concurrency: u16

Expected level of concurrency.

This value controls the number of partitions used internally in the map. A higher value leads to less contention, but also greater memory overhead. The default value is 16.